+++ Character/Player/Goblin.cc
@@ -24,7 +24,7 @@ void Goblin::attackIt(std::shared_ptr<Halfling> e){
         update_message("(");
         update_message(std::to_string(static_cast<int>(e->getHP())));
         update_message("/");
-        update_message(std::to_string(e->getMaxHP()));
+        update_message(std::to_string(static_cast<int>(e->getMaxHP())));
         update_message(")");
     }
     check_dead(e);
@@ -40,7 +40,7 @@ void Goblin::attackIt(std::shared_ptr<Merchant> e){
     update_message("(");
     update_message(std::to_string(static_cast<int>(e->getHP())));
     update_message("/");
-    update_message(std::to_string(e->getMaxHP()));
+    update_message(std::to_string(static_cast<int>(e->getMaxHP())));
     update_message(")");
     e->setNeutral();
     check_dead(e);
