+++ Character/Player/Shade.cc
@@ -23,7 +23,7 @@ void Shade::attackIt(std::shared_ptr<Halfling> e){
         update_message("(");
         update_message(std::to_string(static_cast<int>(e->getHP())));
         update_message("/");
-        update_message(std::to_string(e->getMaxHP()));
+        update_message(std::to_string(static_cast<int>(e->getMaxHP())));
         update_message(").");
     }else{
         update_message("PC misses attack to Halfing. ");
@@ -42,7 +42,7 @@ void Shade::attackIt(std::shared_ptr<Merchant> e){
     update_message("(");
     update_message(std::to_string(static_cast<int>(e->getHP())));
     update_message("/");
-    update_message(std::to_string(e->getMaxHP()));
+    update_message(std::to_string(static_cast<int>(e->getMaxHP())));
     update_message(")");
     e->setNeutral();
     check_dead(e);
